aw-server/requirements.txt
firebase-admin
Flask
Flask-RESTx
aw-core
aw-datastore
aw-query
PyYAML
langchain-google-genai
praisonaiagents
requests
iso8601
Flask-Cors 